+++ README.md
@@ -1,6 +1,6 @@
-[![Build Status](https://img.shields.io/travis/kevinkjt2000/discord-game-server-status/develop.svg?style=flat-square&label=Travis-CI)](https://travis-ci.org/kevinkjt2000/discord-game-server-status?branch=develop)
-[![Coverage Status](https://img.shields.io/coveralls/kevinkjt2000/discord-game-server-status/develop.svg?style=flat-square&label=Coveralls)](https://coveralls.io/github/kevinkjt2000/discord-game-server-status?branch=develop)
-# Game Server Status Bot for Discord
+[![Build Status](https://img.shields.io/travis/kevinkjt2000/bowser/develop.svg?style=flat-square&label=Travis-CI)](https://travis-ci.org/kevinkjt2000/bowser?branch=develop)
+[![Coverage Status](https://img.shields.io/coveralls/kevinkjt2000/bowser/develop.svg?style=flat-square&label=Coveralls)](https://coveralls.io/github/kevinkjt2000/bowser?branch=develop)
+# Bowser Bot for Discord
 Adding this bot to a discord server will allow users to send the message "!status" to see current status about the minecraft server configured in `servers.json`.  Configuration is limited to one minecraft server per discord channel, but you can have a different server on each channel of your discord server if you like.  Eventually I will make my instance of this bot public and add commands that allow some discord role to be able to configure minecraft servers dynamically.  For now, feel free to run the bot, or ping me on discord.
 
 ### Usage
@@ -19,5 +19,5 @@ Adding this bot to a discord server will allow users to send the message "!statu
     }
 }
 ```
-* Run the bot with `python src/main.py`
+* Run the bot with `bowser-bot`
 * Try sending `!help` on the discord server that the bot was added to
