+++ scripts/bowser-bot
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+from bowser.main import main
+
+if __name__ == "__main__":
+    main()
