+++ test/test_Minecraft.py
@@ -1,7 +1,7 @@
 from mcstatus.pinger import PingResponse
 from unittest import TestCase
 from unittest.mock import patch, mock_open, MagicMock
-from src.Minecraft import Minecraft
+from bowser.Minecraft import Minecraft
 from types import SimpleNamespace
 import re
